+++ CHANGELOG.md
@@ -5,9 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0] - 2020-11-09
 ### Added
  - flip function
- - LiftA2 for Applicative
+ - liftA2 for Applicative
  - associative binary operation for Alternative concept
 
 ### Changed